frontend refactor por features
