+++ src/frontend/components/SingUpFormComponent.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import axios from "axios";
 import { useDispatch } from "react-redux";
-import { loginSuccess } from "../../redux/slices/authSlice.js";
+import { loginSuccess } from "../features/auth/authSlice.js";
 import { useNavigate } from "react-router-dom";
 
 export const SignUpFormComponent = () => {
