+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { store } from './redux/store.js'
+import { store } from './app/store'
 import { Provider } from 'react-redux'
 import './index.css'
 import App from './App.jsx'
